+++ test/blutwurst/database_test.clj
@@ -14,7 +14,7 @@
                    :type :complex
                    :properties
                    [{:name "NAME", :type :string, :constraints {:nullable true :maximum-length 100}}
-                    {:name "ID", :type :integer, :constraints {:minimum-value 0 :maximum-value 4294967295 :nullable true}}]})
+                    {:name "ID", :type :integer, :constraints {:minimum-value 0 :maximum-value 2147483647 :nullable true}}]})
 
 (def full-expected-graph {:entities
                           [{:name "PURCHASETYPE" :schema "DBO" :type :complex
@@ -37,12 +37,12 @@
                             [{:name "PURCHASEDBYID",
                               :type :integer,
                               :constraints {:minimum-value 0
-                                            :maximum-value 4294967295,
+                                            :maximum-value 2147483647,
                                             :nullable false}}
                              {:name "PURCHASETYPECATEGORY" :type :string :constraints {:nullable true :maximum-length 50}}
                              {:name "PURCHASETYPENAME" :type :string :constraints {:nullable true :maximum-length 50}}
                              {:name "AMOUNT", :type :decimal,  :constraints {:nullable true :maximum-value 999.99 :minimum-value -999.99}}
-                             {:name "ID", :type :integer,  :constraints {:minimum-value 0 :maximum-value 4294967295 :nullable true}}]}
+                             {:name "ID", :type :integer,  :constraints {:minimum-value 0 :maximum-value 2147483647 :nullable true}}]}
                            person-table]})
 
 (deftest table-graph-tests